Reco_facial/requirements.txt
Django>=3.2,<4.0
djangorestframework
opencv-python
face_recognition
numpy
Pillow
mysqlclient
